% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pts_neighbourhood.R
\name{add_back_nbhds}
\alias{add_back_nbhds}
\title{Add Back Missing Neighbourhoods after \code{onsr::get_pts_neighbourhood()}}
\usage{
add_back_nbhds(data, var, na_to_zero = TRUE)
}
\arguments{
\item{data}{A tibble containing counts of points in neighbourhoods, with fewer than 111 rows. In other words, some neighbourhoods are missing.}

\item{var}{The column containing the count data.}

\item{na_to_zero}{Boolean: Should NA values be converted to 0?}
}
\value{
A 111-row tibble
}
\description{
Add Back Missing Neighbourhoods after \code{onsr::get_pts_neighbourhood()}
}
