% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_geo_types}
\alias{census_geo_types}
\title{Statistics Canada Census GEOID Types}
\format{
A tibble with seven columns:
\describe{
\item{name}{Human-readable name of the GEOUID type.}
\item{abbr}{Vintage geographic reference year abbreviation: either CENSUS, FED, or HRBR.}
\item{type}{A = administrative, S = statistical.}
\item{schema}{A four-digit number that identifies this type of GEOUID.}
\item{geouid}{The type of GEOUID. This is often a column name in StatsCan shapefiles.}
\item{example_name}{Human-readable name of an example geography of this type.}
\item{example_geouid}{API-readable unique GEOUID of the human-readable example geography name.}
...
}
}
\source{
\url{https://www150.statcan.gc.ca/n1/pub/92f0138m/92f0138m2019001-eng.htm}
}
\usage{
census_geo_types
}
\description{
A dataset with descriptions of the different kinds of geographic unique
identifiers (GEOUIDs) Statistics Canada (StatsCan) uses and accepts in its
API calls.
}
\details{
StatsCan defines a number of GEOUID types that correspond to different
geographical units (e.g. provinces, census subdivisions, metropolitan areas).
This table shows all GEOUID types that StatsCan's API accepts, and shows how
to use a specific identifier (e.g. a dissemination block ID) and turn it into
an API-readable GEOUID.

This table is pulled directly from StatsCan, and you can consult StatsCan's
website for more details here:
\itemize{
\item \url{https://www150.statcan.gc.ca/n1/pub/92f0138m/92f0138m2019001-eng.htm}
}
}
\keyword{datasets}
