% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{census_get_geographies}
\alias{census_get_geographies}
\title{Get 2016 Census geographies and geographic attributes for a geographic level.}
\usage{
census_get_geographies(geos = "DA", cpt = 35, lang = "E")
}
\arguments{
\item{geos}{The geographic level to query. Defaults to "DA" for Dissemination
Areas, but the following values are accepted:
\itemize{
\item CD = Census divisions
\item CMACA = Census metropolitan areas and census agglomerations
\item CSD = Census subdivisions (municipalities)
\item CT = Census tracts
\item DA = Dissemination areas
\item DPL = Designated places
\item ER = Economic regions
\item FED = Federal electoral districts (2013 Representation Order)
\item FSA = Forward sortation areas
\item HR = Health regions (including LHINs and PHUs)
\item POPCNTR = Population centres
\item PR = Canada, provinces and territories
}}

\item{cpt}{One province or territory code. Defaults to "35" for Ontario, but
the following values are accepted:
\itemize{
\item 00 = All provinces and territories
\item 10 = Newfoundland and Labrador
\item 11 = Prince Edward Island
\item 12 = Nova Scotia
\item 13 = New Brunswick
\item 24 = Quebec
\item 35 = Ontario
\item 46 = Manitoba
\item 47 = Saskatchewan
\item 48 = Alberta
\item 59 = British Columbia
\item 60 = Yukon
\item 61 = Northwest Territories
\item 62 = Nunavut
}}

\item{lang}{The language in which to return results. There are two options:
\itemize{
\item E = English
\item F = French
}}
}
\value{
A tibble containing parsed API results.
}
\description{
This function provides an interface to Statistics Canada's 2016
Census geographies Web Data Service (WDS) and returns the geographic
identifiers (i.e., DGUIDs) for a geographic level. These DGUIDs are useful
because you can use them to query StatsCan's \emph{other} APIs to retrieve
census data.

Please see StatsCan's website for more information:
\itemize{
\item \url{https://www12.statcan.gc.ca/wds-sdw/cr2016geo-eng.cfm}
}
}
