% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{census_make_dguid}
\alias{census_make_dguid}
\title{Create DGUIDs for Dissemination Areas or Blocks}
\usage{
census_make_dguid(data, type = "DA")
}
\arguments{
\item{data}{A vector or one-column dataframe of DA or DB identifiers.}

\item{type}{The type of census area: accepts "DA" or "DB".}
}
\value{

}
\description{
https://www150.statcan.gc.ca/n1/pub/92f0138m/92f0138m2019001-eng.htm
}
