% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_weighting.R
\name{compute_ons_values}
\alias{compute_ons_values}
\title{Compute standard ONS statistics from a shapefile of points}
\usage{
compute_ons_values(
  point_shapefile,
  name_prefix = "",
  name_postfix = "",
  ons_shp = NA,
  ons_data,
  path_to_od_table,
  verbose = TRUE
)
}
\arguments{
\item{point_shapefile}{A shapefile of points representing service locations.}

\item{name_prefix}{Optional - a prefix to be applied to column names.}

\item{name_postfix}{Optional - a postfix (?) to be applied to column names.}

\item{ons_shp}{Optional. An ONS shapefile. If not specified, uses internal data \code{onsr::ons_shp}.}

\item{ons_data}{ONS data. Can be acquired using \code{onsr::get_ons_data()}.}

\item{path_to_od_table}{File path to an OD table.}

\item{verbose}{Boolean, defaults TRUE. Would you like many updates?}
}
\value{
A tbl_df with the calculated statistics.
}
\description{
This function takes a shapefile of points that can represent any services--
for example, physicians, WIFI stations, or schools-- and some standard ONS
data inputs, and calculates five common ONS statistics:
}
\details{
\itemize{
\item Points per Neighbourhood
\item Points / 1000 residents in the neighbourhood plus a 50m buffer
\item Average travel distance to 3 nearest points
\item Average travel distance to 1 nearest point
\item \% of residents within a 15-minute walk of any point
}

It needs as input a path to a long origin-destination (OD) table such as
can be generated using the Valhalla routing enging and \code{valhallr::od_table()}.
}
\examples{
\dontrun{
# file `wifi_walk_table.csv` computed previously using `valhallr::od_table()`
ons_data <- onsr::get_ons_data()
wifi_data <- onsr::compute_ons_values(point_shapefile = wifi_locations_nad, name_postfix = "wifi", ons_data = ons_data,
                                      path_to_od_table =  "outputs/wifi_walk_table.csv")
}
}
