% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{da_to_ons_data}
\alias{da_to_ons_data}
\title{Linking StatsCan Dissemination Areas to ONS Neighbourhoods}
\format{
A tibble with two columns:
\describe{
\item{DAUID}{Each dissemination area's unique identifier.}
\item{ONS_ID}{A numeric ID corresponding to an ONS neighbourhood.}
...
}
}
\usage{
da_to_ons_data
}
\description{
Linking StatsCan Dissemination Areas to ONS Neighbourhoods
}
\keyword{datasets}
