% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-data.R
\docType{data}
\name{dataset_size}
\alias{dataset_size}
\title{The Dataset with a size column}
\format{
An object of class \code{data.frame} with 41 rows and 3 columns.
}
\usage{
dataset_size
}
\description{
The Dataset with a size column
}
\keyword{internal}
