% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_to_ons_data}
\alias{db_to_ons_data}
\title{Linking StatsCan Dissemination Blocks to ONS Neighbourhoods}
\format{
A tibble with two columns:
\describe{
\item{DBUID}{Each dissemination block's unique identifier.}
\item{ONS_ID}{A numeric ID corresponding to an ONS neighbourhood.}
...
}
}
\usage{
db_to_ons_data
}
\description{
Linking StatsCan Dissemination Blocks to ONS Neighbourhoods
}
\keyword{datasets}
