% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pts_neighbourhood.R
\name{get_db_to_ons}
\alias{get_db_to_ons}
\title{Single-Link Indicator from Dissemination Blocks (DBs) to ONS Neighbourhoods}
\usage{
get_db_to_ons()
}
\value{
A tibble matching each Ottawa DB to one ONS neighbourhood.
}
\description{
Census data is often available at the DB level, whereas ONS
analyses generally take place at the neighbourhood level. This function
returns an SLI that maps each DB in Ottawa to the single ONS neighbourhood
it overlaps the most. The SLI was generated with a simple spatial join from
StatsCan's official DB shapefile and the ONS neighbourhood boundaries,
using the function sf::st_join(largest = TRUE).

See also \code{onsr::get_da_to_ons()}, which maps dissemination areas to ONS
neighbourhoods.
}
