% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ons_shp.R
\name{get_ons_data}
\alias{get_ons_data}
\title{Get ONS Neighbourhood Data}
\usage{
get_ons_data(tidy = TRUE)
}
\arguments{
\item{tidy}{The "tidy" flag determines whether it's returned as a long tibble (default) or a wide one.}
}
\value{
A tibble with all ONS data, including metadata.
}
\description{
Get ONS Neighbourhood Data
}
\examples{
\dontrun{ons_data <- get_ons_data()}
}
