% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ons_shp.R
\name{get_ons_shp}
\alias{get_ons_shp}
\title{Load ONS Gen 2 shapefile}
\usage{
get_ons_shp(all_details = FALSE)
}
\arguments{
\item{all_details}{Boolean. Default is FALSE. If TRUE, returns extra data about each
neighbourhood. Usually not required.}
}
\value{
Simple feature collection with ONS ID number, English name, and French name for 111 ONS Gen2 neighbourhoods.
}
\description{
Load and validate the ONS Gen 2 shapefile from the City of Ottawa's "Open Ottawa" server.
}
\examples{
\dontrun{ons_shp <- get_ons_shp()}
}
