% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_weighting.R
\name{get_pct_within_traveltime}
\alias{get_pct_within_traveltime}
\title{Get percentage of residents within a certain travel time of facilities}
\usage{
get_pct_within_traveltime(walk_table, minute_threshold = 15)
}
\arguments{
\item{walk_table}{A tbl_df with at minimum the columns \code{DBUID} and \code{time},
giving travel time in seconds.}

\item{minute_threshold}{Threshold in minutes. Defaults to 15.}
}
\value{
A tbl_df with columns \code{ONS_ID} and \code{pct_covered}
}
\description{
This function takes a long origin-destination (OD) table between dissemination
blocks (DBs) and facilities and computes, for each ONS neighbourhood and
Ottawa as a whole, the percent of residents within a given travel distance of
those amenities.
}
\details{
OD tables can be generated using the Valhalla routing system and the R
package valhallr.
}
