% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pts_neighbourhood.R
\name{get_pts_neighbourhood}
\alias{get_pts_neighbourhood}
\title{Locate points within ONS neighbourhoods.}
\usage{
get_pts_neighbourhood(pts, pgon = ons_shp)
}
\arguments{
\item{pts}{A simple feature collection class sf containing point data that you would like to locate within polygons (by default, ONS neighbourhoods).}

\item{pgon}{A simple feature collection of class sf containing polygons (by default, ONS neighbourhoods).}
}
\value{
A simple feature collection of class sf with all columns in pts plus additional columns for each point's corresponding polygon (if applicable).
}
\description{
Take an sf geometry object of points and match them with ONS neighbourhoods.
This function is really just a convenience wrapper around a spatial join with useful defaults
input validation. As such, it can also be used with arbitrary polygons.
}
