% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmap_geocode.R
\name{gmap_geocode}
\alias{gmap_geocode}
\title{Tidy Geocoding with Google's API}
\usage{
gmap_geocode(data, var, api_key = NA, verbose = FALSE, wait_nicely = TRUE)
}
\arguments{
\item{data}{A dataframe containing a column with addresses stored as characters.}

\item{var}{A column with addresses stored as characters.}

\item{api_key}{A valid API key for Google Cloud Services. See https://cloud.google.com/free/}

\item{verbose}{Boolean flag controlling console updates. Defaults to FALSE.}

\item{wait_nicely}{Boolean flag controlling API rate. Defaults to TRUE.}
}
\value{
The original dataframe with geocoded latitude and longitude in new columns lat and lng. Addresses that are invalid or not found return NA.
}
\description{
Tidy Geocoding with Google's API
}
