% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmap_geocode.R
\name{gmap_geocode_one}
\alias{gmap_geocode_one}
\title{Geocode a Single Address with Google's API}
\usage{
gmap_geocode_one(
  address = "1243 Willowdale Ave, Ottawa, ON",
  api_key = NA,
  verbose = FALSE,
  wait_nicely = TRUE
)
}
\arguments{
\item{address}{A character vector with a single address.}

\item{api_key}{A valid API key for Google Cloud Services. See https://cloud.google.com/free/}

\item{verbose}{Boolean flag controlling console updates. Defaults to FALSE.}

\item{wait_nicely}{Boolean flag controlling API rate. Defaults to TRUE.}
}
\value{
A dataframe with geocoded latitute and logitude.
}
\description{
Geocode a Single Address with Google's API
}
