% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-browse.R
\name{ons_browse_qmi}
\alias{ons_browse_qmi}
\title{Quickly browse to dataset's QMI}
\usage{
ons_browse_qmi(id = NULL)
}
\arguments{
\item{id}{\verb{[character]}     Id that represents a dataset.}
}
\description{
This function take you to the QMI
}
\examples{
ons_browse_qmi("cpih01")
}
