% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-datasets.R
\name{ons_datasets}
\alias{ons_datasets}
\alias{ons_ids}
\alias{ons_lookout}
\alias{ons_editions}
\title{ONS Datasets}
\usage{
ons_datasets()

ons_ids()

ons_lookout(id = NULL, desc = FALSE)

ons_editions(id = NULL)
}
\description{
Used to find information about data published by the ONS.
}
\examples{


ons_lookout("cpih01")
ons_lookout("ageing-population-estimates")

}
