% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-get.R
\name{ons_extra}
\alias{ons_extra}
\alias{ons_get_dim}
\alias{ons_get_dim_opts}
\alias{ons_get_meta}
\alias{ons_get_obs}
\title{Access dataset's addtiional information.}
\usage{
ons_get_dim(id = NULL, edition = NULL, version = NULL, limit = 20, offset = 0)

ons_get_dim_opts(
  id = NULL,
  edition = NULL,
  version = NULL,
  dimension = NULL,
  limit = 20,
  offset = 0
)

ons_get_meta(id = NULL, edition = NULL, version = NULL)

ons_get_obs(id = NULL, edition = NULL, version = NULL, ...)
}
\description{
Data in each version is broken down by \code{dimensions}, and a unique
combination of dimension \code{options} in a version can be used to retrieve
\code{observation} level data.
}
\examples{
ons_get_dim(id = "cpih01")

ons_get_dim_opts(id = "cpih01", dimension = "time", limit = 400)

ons_get_meta(id = "cpih01")

ons_get(id = "cpih01", edition = "time-series", version = 3)
ons_get(id = "cpih01", edition = "time-series", version = 3, dimension = "geography")


# Take only specific observations
ons_get_obs("cpih01", geography = "K02000001", aggregate = "cpih1dim1A0", time = "Oct-11")

# Or can use a wildcard for the time
ons_get_obs("cpih01", geography = "K02000001", aggregate = "cpih1dim1A0", time = "*")
}
