% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ons_ids}
\alias{ons_ids}
\title{ONS neighbourhood ID numbers.}
\format{
A tibble with 111 rows and one column:

\describe{
\item{ONS_ID}{Unique identifier, integer}
...
}
}
\source{
\url{https://www.neighbourhoodstudy.ca/}
}
\usage{
ons_ids
}
\description{
A single-column tibble containing all ONS IDs. Useful for ensuring that all
ONS_IDs are represented in a given dataset, and for adding back any that are
missing.#'
}
\keyword{datasets}
