% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_weighting.R
\name{ons_pop_weight_dbs}
\alias{ons_pop_weight_dbs}
\title{Get neighbourhood population-weighted travel distances and times from DBs to facilities}
\usage{
ons_pop_weight_dbs(od_table, n_closest = 5, verbose = TRUE)
}
\arguments{
\item{od_table}{A tidy origin-destination table with (at least) columns
\code{DBUID}, \code{distance}, and \code{time}. Output from \code{valhallr::od_table()} should
be accepted seamlessly.}

\item{n_closest}{The number of closest facilities to consider. Default is 5.}

\item{verbose}{Boolean. Would you like lots of updates in the console?}
}
\value{
A 111-row tibble with population-weighted travel distances and times
for each ONS neighbourhood.
}
\description{
This function calculates neighbourhood-level population-weighted
average travel times and distances from each dissemination block (DB) to the
n-nearest facilities. Each DB is linked to the single neighbourhood it
overlaps the most

It works well with the function \code{valhallr::od_table()} from the \strong{valhallr}
package.
}
