% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ons_shp}
\alias{ons_shp}
\title{Details and geographies for Ottawa Neighbourhood Study (ONS) Gen2 boundaries.}
\format{
A simple feature geometry collection with 111 features/rows
and 3 fields/columns:
\describe{
\item{ONS_ID}{Unique identifier, integer}
\item{Name}{English name}
\item{Name_FR}{French name}
\item{geometry}{Neighbourhood boundary, multipolygon geometry}
...
}
}
\source{
\url{https://www.neighbourhoodstudy.ca/}
}
\usage{
ons_shp
}
\description{
A dataset containing the ONS IDs, English and French names, and boundaries
for 111 neighbourhoods in Ottawa.
}
\keyword{datasets}
