% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ottawa_db_pops_2016}
\alias{ottawa_db_pops_2016}
\title{Dissemination Block (DB) Populations, Ottawa, 2016 Census}
\format{
A tibble with two columns:
\describe{
\item{DBUID}{Unique dissemination block identifier}
\item{db_pop_2016}{DB population from 2016 census.}
...
}
}
\source{
\url{https://open.canada.ca/data/en/dataset/32f1a777-9fcf-4e4a-8c66-82c66a2e76f1}
}
\usage{
ottawa_db_pops_2016
}
\description{
A dataset containing DB-level populations for all 8,016 DBs in Ottawa.
}
\details{
Data is derived from the Statistics Canada 2016 Census Geographic Attribute
File, accessible here:
\itemize{
\item \url{https://open.canada.ca/data/en/dataset/32f1a777-9fcf-4e4a-8c66-82c66a2e76f1}
}
}
\keyword{datasets}
