% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{print_md_table}
\alias{print_md_table}
\title{Print data in a tibble to the console as a Markdown table}
\usage{
print_md_table(mydata)
}
\arguments{
\item{mydata}{A tibble/dataframe to format as a Markdown table.}
}
\value{
The original data.
}
\description{
Print data in a tibble to the console as a Markdown table
}
