% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_scraping.R
\name{scrape_yelp}
\alias{scrape_yelp}
\title{Get Business Names & Addresses from Yelp}
\usage{
scrape_yelp(search_term, location = "Ottawa, ON", num_pages = 24)
}
\arguments{
\item{search_term}{Character string to search Yelp for.}

\item{location}{Character string containing the location to search. Defaults
to "Ottawa, ON".}

\item{num_pages}{The number of pages to search. Defaults to 24.}
}
\value{
A tibble with Yelp search results.
}
\description{
To support the Ottawa Neighbourhood Study in refreshing its food availability
database, this function does a perfunctory Yelp search for a given term in a
given location.
}
\details{
Results include business names and addresses. This function could be improved
but it works for what we need now, and Yelp is very protective of its data.
Be warned that scraping too much too quickly may get your IP address blocked.
}
\examples{
\dontrun{
convenience_stores <- scrape_yelp("convenience stores")
}
}
