% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_md_table}
\alias{write_md_table}
\title{Save data in a tibble to file as a Markdown table}
\usage{
write_md_table(mydata, destfile = NA)
}
\arguments{
\item{mydata}{A tibble/dataframe to format as a Markdown table.}

\item{destfile}{File where output will be saved.}
}
\value{
The original data.
}
\description{
Save data in a tibble to file as a Markdown table
}
