% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{census_get_data}
\alias{census_get_data}
\title{Get StatsCan Census Data for Specified Regions}
\usage{
census_get_data(dguids = NA, topic = 0, notes = 0, stat = 0, lang = "E")
}
\arguments{
\item{dguids}{A character vector or one-column dataframe containing the
DGUIDs to query. \emph{Note: in a future version this will be made "tidier."}}

\item{topic}{Which subset of census data should be returned? The following values are accepted;
\itemize{
\item 0 = All topics (default)
\item 1 = Aboriginal peoples
\item 2 = Education
\item 3 = Ethnic origin
\item 4 = Families, households and marital status
\item 5 = Housing
\item 6 = Immigration and citizenship
\item 7 = Income
\item 8 = Journey to work
\item 9 = Labour
\item 10 = Language
\item 11 = Language of work
\item 12 = Mobility
\item 13 = Population
\item 14 = Visible minority
}}

\item{notes}{Should footnotes be returned? 0 = no (default), 1 = yes.}

\item{stat}{What statistic should be returned? 0 = counts (default), 1 = rates.}
}
\value{

}
\description{
This function provides an interface to Statistics Canada's 2016
Census geographies Web Data Service (WDS) and returns some or all Census
Profile data for a geography of interest. Data is returned for individual
regions specified by Dissemination Geography Unique Identifiers (or DGUIDs),
which can correspond to census regions like dissemination areas, political
regions like provinces and territories, or economic regions.

\strong{NOTE:} This function is under active development, and it is my intention
to turn it into a properly "tidy" function.

For more details on the underlying API, please see StatsCan's documentation here:
\itemize{
\item \url{https://www12.statcan.gc.ca/wds-sdw/cpr2016-eng.cfm}
}
}
\details{
\emph{Note: This function is under active development. Please send suggestions or
bug reports to christopher.a.belanger (at) gmail.com.}
}
\examples{
\dontrun{
# get mobility data for Ottawa and Leeds & Grenville, Census Divisions 3506 &
3507.
onsr::census_make_dguid(data = c("3506", "3507"),
                       geouid_type = "CDUID",
                       vintage_year = 2016) \%>\%
onsr:::census_get_data(topic = 12)
}
}
