% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{geocode_ottawa}
\alias{geocode_ottawa}
\title{Tidy Geocoding with the City of Ottawa's API}
\usage{
geocode_ottawa(data, var, verbose = FALSE, wait_nicely = TRUE)
}
\arguments{
\item{data}{A dataframe containing a column with addresses stored as characters.}

\item{var}{A column with addresses stored as characters.}

\item{verbose}{Boolean flag controlling console updates. Defaults to FALSE.}

\item{wait_nicely}{Boolean flag controlling API rate. Defaults to TRUE.}
}
\value{
The original dataframe with geocoded latitude and longitude in new columns lat and lng. Addresses that are invalid or not found return NA.
}
\description{
Tidy Geocoding with the City of Ottawa's API
}
