% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-browse.R
\name{ons_browse}
\alias{ons_browse}
\title{Quickly browse to ONS' developer webpage}
\usage{
ons_browse()
}
\value{
The url of the webpage
}
\description{
This function take you to the ONS' developer webpage.
}
\examples{
\donttest{
ons_browse()
}
}
