% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-code-lists.R
\name{ons_codelists}
\alias{ons_codelists}
\alias{ons_codelist}
\title{Explore codes and lists}
\usage{
ons_codelists()

ons_codelist(id = NULL)
}
\description{
Used to get details about codes and code lists stored by ONS. Codes are used
to provide a common definition when presenting statistics with related categories.
Codes are gathered in code lists, which may change over time to include new or
different codes. The meaning of a code should not change over time, but new codes
may be created where new meaning is required.
}
\examples{

ons_codelists()
ons_codelist(id = "quarter")

#editions
ons_codelist_editions(id = "quarter")
ons_codelist_edition(id = "quarter", edition = )

#codes
ons_codes(id = "quarter", edition = "one-off")
ons_code(id = "quarter", edition = "one-off", code = "q2")

ons_code_dataset(id = "quarter", edition = "one-off", code = "q2")

}
