% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-code-lists.R
\name{ons_codelists}
\alias{ons_codelists}
\alias{ons_codelist}
\alias{ons_codelist2}
\alias{ons_codelist3}
\alias{ons_codelist4}
\title{Explore codes and lists}
\usage{
ons_codelists()

ons_codelist(id = NULL, edition = NULL, code = NULL)

ons_codelist2(id = NULL, edition = NULL, code = NULL)

ons_codelist3(id = NULL, edition = NULL, code = NULL)

ons_codelist4(id = NULL, edition = NULL, code = NULL)
}
\description{
Used to get details about codes and code lists stored by ONS. Codes are used
to provide a common definition when presenting statistics with related categories.
Codes are gathered in code lists, which may change over time to include new or
different codes. The meaning of a code should not change over time, but new codes
may be created where new meaning is required.
}
\examples{

ons_codelists()
ons_codelist(id = "quarter")

#editions
ons_codelist2(id = "quarter", edition = EMPTY)
ons_codelist2(id = "quarter", edition = "one-off")

#codes
ons_codelist3(id = "quarter", edition = "one-off", code = EMPTY)
ons_codelist3(id = "quarter", edition = "one-off", code = "q2")

ons_codelist4(id = "quarter", edition = "one-off", code = "q2")

}
