% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-code-lists.R
\name{ons_codes}
\alias{ons_codes}
\alias{ons_code}
\alias{ons_code_dataset}
\title{Explore codes and lists}
\usage{
ons_codes(id = NULL, edition = NULL)

ons_code(id = NULL, edition = NULL, code = NULL)

ons_code_dataset(id = NULL, edition = NULL, code = NULL)
}
\arguments{
\item{id}{\verb{[character]}     Id that represents a dataset.}

\item{edition}{\verb{[character]}     A subset of the dataset representing a specific time period.
For some datasets this edition can contain all time periods (all historical data).
The latest version of this is displayed by default.}

\item{code}{\verb{[character]} The ID of the code within a code list.}
}
\description{
Used to get details about codes and code lists stored by ONS. Codes are used
to provide a common definition when presenting statistics with related categories.
Codes are gathered in code lists, which may change over time to include new or
different codes. The meaning of a code should not change over time, but new codes
may be created where new meaning is required.
}
\examples{


#codes
ons_codes(id = "quarter", edition = "one-off")
ons_code(id = "quarter", edition = "one-off", code = "q2")

ons_code_dataset(id = "quarter", edition = "one-off", code = "q2")

}
