% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-datasets.R, R/02-get.R
\name{ons_datasets}
\alias{ons_datasets}
\alias{ons_ids}
\alias{ons_get_dims}
\alias{ons_get_dim_opts}
\alias{ons_get_meta}
\alias{ons_get_obs}
\title{ONS Datasets}
\usage{
ons_datasets()

ons_ids()

ons_get_dims(id = NULL, edition = NULL, version = NULL)

ons_get_dim_opts(id = NULL, edition = NULL, version = NULL, dimension = NULL)

ons_get_meta(id = NULL, edition = NULL, version = NULL)

ons_get_obs(id = NULL, edition = NULL, version = NULL)
}
\description{
Used to find information about data published by the ONS.
}
\examples{

ons_dataset(id = "ageing-population-estimates", edition = "time-series", version = 1)
ons_dataset(id = "ageing-population-estimates")

ons_dataset_dims(id = "cpih01", edition = "time-series", version = 3)


ons_dataset_dims(id = "cpih01", edition = "time-series", version = 3)
ons_dataset_dim_opts(id = "cpih01", edition = "time-series", version = 3, dimension = "geography")

}
