% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-datasets.R
\name{ons_datasets}
\alias{ons_datasets}
\alias{ons_ids}
\title{ONS Datasets}
\usage{
ons_datasets()

ons_ids()
}
\value{
A tibble with the datasets.
}
\description{
A grouping of data (editions) with shared dimensions, for example Sex,
Age and Geography, and all published history of this group of data.
The options in these dimensions can change over time leading to
separate editions. For example: Population Estimates for UK, England
and Wales, Scotland and Northern Ireland.
}
\examples{
\donttest{
# Find all the information about the data
ons_datasets()

# Just the ids
ons_ids()
}
}
