% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-get.R
\name{ons_get}
\alias{ons_get}
\alias{ons_get_dim_opts}
\alias{ons_get_meta}
\alias{ons_get_obs}
\title{Get data from}
\usage{
ons_get(
  id = NULL,
  edition = NULL,
  version = NULL,
  ons_read = getOption("onsr.read"),
  ...
)

ons_get_dim_opts(id = NULL, edition = NULL, version = NULL, dimension = NULL)

ons_get_meta(id = NULL, edition = NULL, version = NULL)

ons_get_obs(id = NULL, edition = NULL, version = NULL)
}
\arguments{
\item{id}{`[character]`}

\item{edition}{`[character]`}

\item{version}{`[version]`}

\item{ons_read}{`[ons_read]`}

\item{...}{Further arguments passed on the reading functions.}
}
\description{
Get data from
}
\examples{
ons_get(id = "cpih01", edition = "time-series", version = 5)

ons_get(id = "cpih01", edition = "time-series")

ons_get(id = "ageing-population-estimates", edition = "time-series", version = 1)
ons_get(id = "ageing-population-estimates")

ons_get_dims(id = "cpih01", edition = "time-series", version = 3)


ons_dataset_dims(id = "cpih01", edition = "time-series", version = 3)
ons_dataset_dim_opts(id = "cpih01", edition = "time-series", version = 3, dimension = "geography")



}
